Modularized Spotify API calls
