+++ .prettierignore
@@ -1,3 +1,4 @@
 .github
 README.md
 node_modules
+dist
